pyproject.toml
[tool.poetry]
name = "llamaindex-bottoms-up-development"
version = "0.1.0"
description = ""
authors = ["Achilleas Voutsas <achilleas.voutsas@ema.europa.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sentence-transformers = "2.2.2"
torch = "^2.3.1"
python-dotenv = "^1.0.1"
instructorembedding = "^1.0.1"
llama-index = "^0.10.43"
llama-index-embeddings-openai = "^0.1.10"
openai = "^1.33.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
mypy = "^1.10.0"
wheel = "^0.43.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
